[Feat] 창고 UI 구현
