--- Assets/SHG/Scripts/UI/InventoryItemContainerWindow.cs
@@ -1,97 +1,0 @@
-using System;
-using System.Collections.Generic;
-using UnityEngine.UIElements;
-
-namespace SHG
-{
-  public class InventoryItemContainerWindow : ItemConatinerWindow
-  {
-    const MouseButton DRAG_BUTTON = ItemConatinerWindow.MouseButton.Left;
-    const MouseButton USE_BUTTON = ItemConatinerWindow.MouseButton.Right;
-    Dictionary<VisualElement, ItemAndCount> itemBoxTable;
-    Func<ItemData, bool> filterItem;
-
-    public InventoryItemContainerWindow(Func<ItemData, bool> filterItem, ItemBox floatingItemBox): base (floatingItemBox)
-    {
-      this.itemBoxTable = new ();
-      this.filterItem = filterItem;
-    }
-
-    protected override void CreateUI()
-    {
-      var label = new Label();
-      label.text = this.label;
-      label.AddToClassList("window-label");
-      this.Add(label);
-    }
-
-    //TODO: 각 아이템 UI를 objectpool에 보관
-    ItemBox CreateItembox(ItemAndCount itemAndCount)
-    {
-      ItemBox itemBox = new ItemBox(this);
-      itemBox.SetData(itemAndCount);
-      itemBox.RegisterCallback<PointerDownEvent>(this.OnPointerDown);         
-      itemBox.RegisterCallback<PointerUpEvent>(this.OnPointerUp);
-      itemBox.RegisterCallback<PointerMoveEvent>(this.OnPointerMove);
-      this.itemBoxTable[itemBox] = itemAndCount;
-      return (itemBox);
-    }
-
-    protected override void FillItems(ItemStorageBase inventory)
-    {
-      foreach (var pair in inventory.Items) {
-        var (item, count) = pair;
-        if (count > 0 && this.filterItem(item)) {
-          var box = this.CreateItembox(
-            new ItemAndCount { Item = item, Count = count });
-          this.itemsContainer.Add(box);
-        }
-      } 
-    }
-
-    // TODO: return to Obejctpool
-    protected override void ClearItems()
-    {
-      this.itemsContainer.Clear();
-    }
-
-    protected override void OnUsePointerButtonDown(ItemBox boxElement, ItemAndCount itemAndCount)
-    {
-      var item = App.Instance.Inventory.PeakItem(itemAndCount.Item); 
-      if (item is IUsable usableItem) {
-        this.UseItem(App.Instance.Inventory.GetItem(itemAndCount.Item) as IUsable);
-      }
-    }
-
-    void UseItem(IUsable usableItem)
-    {
-      App.Instance.Inventory.UseItem(usableItem);
-    }
-
-    void OnDropItemToQuickSlot(EquipmentItemData equipmentItemData)
-    {
-      App.Instance.Inventory.MoveItemToQuickSlot(equipmentItemData);
-    }
-
-    protected override bool IsAbleToDropItem(ItemData item)
-    {
-      return (item as EquipmentItemData);
-    }
-
-    protected override void DropItem(ItemAndCount itemAndCount)
-    {
-      App.Instance.Inventory.MoveItemToQuickSlot(itemAndCount.Item);
-    }
-
-    protected override void DropItemOutSide(ItemAndCount itemAndCount)
-    {
-      //FIXME: drop item count?
-      App.Instance.Inventory.GetItem(itemAndCount.Item);
-    }
-
-    protected override bool IsAbleToDropOutSide(ItemData item)
-    {
-      return (true);
-    }
-  }
-}
