+++ Assets/SHG/Scripts/UI/InventoryWindow.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3939bb10c087b3aeebd1319b5fc3fa35
+guid: 348d98842c96c9ac38f73525475487f4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
