+++ Assets/SHG/Scripts/UI/ItemContainerStyle.uss
