+++ Assets/SHG/Scripts/UI/ItemStorageWindow.cs
@@ -6,9 +6,9 @@ using UnityEngine.UIElements;
 
 namespace SHG
 {
-  public class ItemStorageInnerWindow : ItemConatinerWindow
+  public class ItemStorageWindow : ItemConatinerWindow
   {
-    public ItemStorageInnerWindow(ItemBox floatingItemBox) : base(floatingItemBox)
+    public ItemStorageWindow(ItemBox floatingItemBox) : base(floatingItemBox, App.Instance.ItemStorage)
     {
     }
 
@@ -19,12 +19,21 @@ namespace SHG
 
     protected override void CreateUI()
     {
-      throw new NotImplementedException();
     }
 
-    protected override void DropItem(ItemAndCount itemAndCount)
+    protected override void DropItem(ItemAndCount itemAndCount, ItemConatinerWindow targetContainer)
     {
-      throw new NotImplementedException();
+      if (targetContainer is InventoryWindow inventoryItemContainer) {
+
+        if (itemAndCount.Count > 1) {
+          var (item, count) = this.ItemSource.GetItems(itemAndCount.Item, itemAndCount.Count);
+          targetContainer.ItemSource.AddItems(item, count);
+        }
+        else {
+          var item = this.ItemSource.GetItem(itemAndCount.Item);
+          targetContainer.ItemSource.AddItem(item);
+        }
+      }
     }
 
     protected override void DropItemOutSide(ItemAndCount itemAndCount)
@@ -34,32 +43,52 @@ namespace SHG
 
     protected override void FillItems(ItemStorageBase inventory)
     {
-      throw new NotImplementedException();
+      foreach (var itemAndCount in inventory.Items) {
+        var (item, count) = itemAndCount;
+        if (count > 0) {
+          var box = this.CreateItembox(
+            new ItemAndCount { Item = item, Count = count });
+          this.itemsContainer.Add(box);
+        }
+      }
+    }
+    //TODO: 각 아이템 UI를 objectpool에 보관
+    ItemBox CreateItembox(ItemAndCount itemAndCount)
+    {
+      ItemBox itemBox = new ItemBox(this);
+      itemBox.SetData(itemAndCount);
+      itemBox.RegisterCallback<PointerDownEvent>(this.OnPointerDown);         
+      itemBox.RegisterCallback<PointerUpEvent>(this.OnPointerUp);
+      itemBox.RegisterCallback<PointerMoveEvent>(this.OnPointerMove);
+      return (itemBox);
     }
 
-    protected override bool IsAbleToDropItem(ItemData item)
+    protected override bool IsAbleToDropItem(ItemData item, ItemConatinerWindow targetContainer)
     {
-      throw new NotImplementedException();
+      if (targetContainer is InventoryWindow inventoryItemContainer) {
+        return (true);
+      }
+      return (false);
     }
 
     protected override bool IsAbleToDropOutSide(ItemData item)
     {
-      throw new NotImplementedException();
+      return (false);
     }
 
     protected override void OnUsePointerButtonDown(ItemBox boxElement, ItemAndCount itemAndCount)
     {
       throw new NotImplementedException();
     }
-    }
+  }
 
-    public class ItemStorageWindow : VisualElement, IHideableUI
+  public class ItemStorageContainerWindow : VisualElement, IHideableUI
   {
     public bool IsVisiable { get; private set; }
-    ItemStorageInnerWindow ItemContainer;
+    public ItemStorageWindow ItemContainer { get; private set; }
     ItemBox floatingBox;
 
-    public ItemStorageWindow(ItemBox floatingBox)
+    public ItemStorageContainerWindow(ItemBox floatingBox)
     {
       this.floatingBox = floatingBox;
       this.name = "item-storage-window-container";
@@ -71,12 +100,14 @@ namespace SHG
     {
       this.IsVisiable = false;
       Utils.HideVisualElement(this);
+      this.ItemContainer.Hide();
     }
 
     public void Show()
     {
       this.IsVisiable = true;
       Utils.ShowVisualElement(this);
+      this.ItemContainer.Show(); 
     }
 
     void CreateUI()
@@ -90,8 +121,8 @@ namespace SHG
       closeButton.AddToClassList("window-close-button");
       closeButton.RegisterCallback<ClickEvent>(this.OnClickClose);
       this.Add(closeButton); 
-      this.ItemContainer = new ItemStorageInnerWindow(
-          this.floatingBox
+      this.ItemContainer = new ItemStorageWindow(
+        this.floatingBox
         );
       this.ItemContainer.name = "item-storage-item-container";
       this.Add(this.ItemContainer);
