+++ Assets/SHG/Scripts/UI/QuickSlotStyle.uss
