+++ Assets/SHG/Scripts/UI/QuickSlotStyle.uss.meta
